backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
# psycopg2-binary==2.9.9 # Fails on Python 3.14
psycopg==3.1.18
gunicorn==21.2.0
Flask-Migrate==4.0.5
pandas==3.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
openpyxl==3.1.5
